+++ GalaxyEngine/src/Render/Camera.cpp
@@ -67,7 +67,7 @@ namespace GALAXY {
 
 	Mat4 Render::Camera::GetProjectionMatrix()
 	{
-		float tanHalfFov = std::tanf(p_fov * DegToRad * 0.5f);
+		float tanHalfFov = std::tan(p_fov * DegToRad * 0.5f);
 
 		Mat4 projectionMatrix = Mat4();
 		projectionMatrix[0][0] = 1.0f / (p_aspectRatio * tanHalfFov);
